# Separated the logic of branch and bound search