--- a/REQUIRE
+++ b/REQUIRE
@@ -1,4 +1,4 @@
-julia 0.7
+julia 1.0
 IntervalArithmetic 0.15
 ForwardDiff 0.8
 StaticArrays 0.8