--- a/src/IntervalRootFinding.jl
+++ b/src/IntervalRootFinding.jl
@@ -46,7 +46,7 @@ include("krawczyk.jl")
 
 include("complex.jl")
 include("contractors.jl")
-include("rootsearch_iterator.jl")
+include("branch_and_bound.jl")
 include("roots.jl")
 include("newton1d.jl")
 include("quadratic.jl")