--- a/src/branch_and_bound.jl
+++ b/src/branch_and_bound.jl
@@ -35,7 +35,7 @@ function BBLeaf(data::DATA, parent::Int) where {DATA}
 end
 
 function BBNode(leaf::BBLeaf, child1::Int, child2::Int)
-    BBNode(parent_id(leaf), Int[child1, child2])
+    BBNode(leaf.parent, Int[child1, child2])
 end
 
 """
@@ -49,13 +49,14 @@ Support the iterator interface. The element yielded by the iteration are
 tuples `(node_id, lvl)` where `lvl` is the depth of the node in the tree.
 """
 struct BBTree{DATA}
-    nodes::Dict{Int, Union{BBNode, BBLeaf{DATA}}}
-    working_leafs::Vector{Int}
+    nodes::Dict{Int, BBNode}
+    leaves::Dict{Int, BBLeaf{DATA}}
+    working_leaves::Vector{Int}
 end
 
 function BBTree(rootdata::DATA) where {DATA}
     rootleaf = BBLeaf(rootdata, 0)
-    BBTree{DATA}(Dict{Int, Union{BBNode, BBLeaf{DATA}}}(1 => rootleaf), Int[1])
+    BBTree{DATA}(Dict{Int, BBNode}(), Dict(1 => rootleaf), Int[1])
 end
 
 show(io::IO, wn::BBNode) = print(io, "Node with children $(wn.children)")
@@ -67,7 +68,7 @@ function show(io::IO, wt::BBTree{DATA}) where {DATA}
     println(io, "Working tree with $(nnodes(wt)) elements of type $DATA")
 
     if nnodes(wt) > 0
-        println(io, "Indices: ", keys(wt.nodes) |> collect |> sort)
+        println(io, "Indices: ", vcat(collect(keys(wt.nodes)), collect(keys(wt.leaves))) |> sort)
         println(io, "Structure:")
         for (id, lvl) in wt
             println(io, "  "^lvl * "[$id] $(wt[id])")
@@ -75,27 +76,53 @@ function show(io::IO, wt::BBTree{DATA}) where {DATA}
     end
 end
 
-parent_id(node::AbstractBBNode) = node.parent
-parent_id(wt::BBTree, node::AbstractBBNode) = parent_id(node)
-parent_id(wt::BBTree, id::Int) = parent_id(wt[id])
-parent(wt::BBTree, node::AbstractBBNode) = wt[parent_id(node)]
-parent(wt::BBTree, id::Int) = wt[parent_id(wt, id)]
-children_ids(node::BBNode) = node.children
-children_ids(wt::BBTree, id::Int) = children_ids(wt[id])
 data(leaf::BBLeaf) = leaf.data
 
 # Root node has id 1 and parent id 0
 root(wt::BBTree) = wt[1]
 is_root(wt::BBTree, id::Int) = (id == 1)
 
-nnodes(wt::BBTree) = length(wt.nodes)
-newid(wt::BBTree) = maximum(keys(wt.nodes)) + 1
-data(wt::BBTree) = [data(val) for val in values(wt.nodes) if isa(val, BBLeaf)]
+nnodes(wt::BBTree) = length(wt.nodes) + length(wt.leaves)
+data(wt::BBTree) = data.(values(wt.leaves))
 
-# Index operations
-getindex(wt::BBTree, id) = wt.nodes[id]
-setindex!(wt::BBTree, id, val) = setindex!(wt.nodes, id, val)
-delete!(wt::BBTree, id) = delete!(wt.nodes, id)
+function newid(wt::BBTree)
+    k1 = keys(wt.nodes)
+    k2 = keys(wt.leaves)
+
+    if length(k1) > 0
+        m1 = maximum(k1)
+    else
+        m1 = 0
+    end
+
+    if length(k2) > 0
+        m2 = maximum(k2)
+    else
+        m2 = 0
+    end
+
+    return max(m1, m2) + 1
+end
+
+# Index operations (probably slow)
+function getindex(wt::BBTree, id)
+    haskey(wt.nodes, id) && return wt.nodes[id]
+    haskey(wt.leaves, id) && return wt.leaves[id]
+    error("getindex failed: no index $id")  # TODO: make better error
+end
+
+setindex!(wt::BBTree, val::BBNode, id) = setindex!(wt.nodes, val, id)
+setindex!(wt::BBTree, val::BBLeaf, id) = setindex!(wt.leaves, val, id)
+
+function delete!(wt::BBTree, id)
+    if haskey(wt.nodes, id)
+        delete!(wt.nodes, id)
+    elseif haskey(wt.leaves, id)
+        delete!(wt.leaves, id)
+    else
+        error("delete! failed: no index $id")  # TODO: make better error
+    end
+end
 
 """
     discard_leaf!(wt::BBTree, id::Int)
@@ -104,20 +131,22 @@ Delete the `BBLeaf` with index `id` and all its ancestors to which it is
 the last descendant.
 """
 function discard_leaf!(wt::BBTree, id::Int)
-    leaf = wt[id]
-    recursively_delete_child!(wt, parent_id(leaf), id)
+    leaf = wt.leaves[id]
+    delete!(wt.leaves, id)
+    recursively_delete_parent!(wt, leaf.parent, id)
 end
 
-function recursively_delete_child!(wt, id_parent, id_child)
+function recursively_delete_parent!(wt, id_parent, id_child)
     if !is_root(wt, id_child)
-        parent = wt[id_parent]
-        cc = children_ids(parent)
-        filter!(id -> id != id_child, cc)
-        if isempty(cc)
-            recursively_delete_child!(wt, parent_id(parent), id_parent)
+        parent = wt.nodes[id_parent]
+        siblings = parent.children
+        if length(parent.children) == 1  # The child has no siblings, so delete the parent
+            delete!(wt.nodes, id_parent)
+            recursively_delete_parent!(wt, parent.parent, id_parent)
+        else  # The child has siblings so remove it from the children list
+            deleteat!(parent.children, searchsortedfirst(parent.children, id_child))
         end
     end
-    delete!(wt, id_child)
 end
 
 function iterate(wt::BBTree, (id, lvl)=(0, 0))
@@ -129,14 +158,14 @@ end
 function next_id(wt::BBTree, id, lvl)
     lvl == 0 && return (1, 1)
     node = wt[id]
-    isa(node, BBNode) && return (children_ids(node)[1], lvl + 1)
+    isa(node, BBNode) && return (node.children[1], lvl + 1)
     return next_sibling(wt, id, lvl)
 end
 
 function next_sibling(wt::BBTree, sibling, lvl)
-    parent = parent_id(wt, sibling)
+    parent = wt[sibling].parent
     parent == 0 && return (0, 0)
-    children = children_ids(wt, parent)
+    children = wt[parent].children
     maximum(children) == sibling && return next_sibling(wt, parent, lvl - 1)
     id = minimum(filter(x -> x > sibling, children))
     return (id, lvl)
@@ -187,38 +216,38 @@ abstract type KeyBBSearch{DATA} <: BBSearch{DATA} end
     get_leaf_id!(::BBSearch, wt::BBTree)
 
 Return the id of the next leaf that will be processed and remove it from the
-list of working leafs.
+list of working leaves.
 
 Must be define for custom searches that are direct subtype of `BBSearch`.
 """
-get_leaf_id!(::BreadthFirstBBSearch, wt::BBTree) = popfirst!(wt.working_leafs)
-get_leaf_id!(::DepthFirstBBSearch, wt::BBTree) = pop!(wt.working_leafs)
-get_leaf_id!(::KeyBBSearch, wt::BBTree) = popfirst!(wt.working_leafs)
+get_leaf_id!(::BreadthFirstBBSearch, wt::BBTree) = popfirst!(wt.working_leaves)
+get_leaf_id!(::DepthFirstBBSearch, wt::BBTree) = pop!(wt.working_leaves)
+get_leaf_id!(::KeyBBSearch, wt::BBTree) = popfirst!(wt.working_leaves)
 
 """
     insert_leaf!(::BBSearch, wt::BBTree, leaf::BBLeaf)
 
 Insert the id of a new leaf that has been produced by bisecting an older leaf
-into the list of working leafs.
+into the list of working leaves.
 
 Must be define for custom searches that are direct subtype of `BBSearch`.
 """
 function insert_leaf!(::Union{BreadthFirstBBSearch{DATA}, DepthFirstBBSearch{DATA}},
                       wt::BBTree{DATA}, leaf::BBLeaf{DATA}) where {DATA}
     id = newid(wt)
-    wt[id] = leaf
-    push!(wt.working_leafs, id)
+    wt.leaves[id] = leaf
+    push!(wt.working_leaves, id)
     return id
 end
 
 function insert_leaf!(::KS, wt::BBTree{DATA}, leaf::BBLeaf{DATA}) where {DATA, KS <: KeyBBSearch{DATA}}
     id = newid(wt)
-    wt[id] = leaf
-    keys = keyfunc.(KS, wt.working_leafs)
+    wt.leaves[id] = leaf
+    keys = keyfunc.(KS, wt.working_leaves)
     current = keyfunc(KS, leaf)
 
-    # Keep the working_leafs sorted
-    insert!(wt.working_leafs, searchsortedfirst(keys, current), id)
+    # Keep the working_leaves sorted
+    insert!(wt.working_leaves, searchsortedfirst(keys, current), id)
     return id
 end
 
@@ -227,25 +256,26 @@ IteratorSize(::Type{BBS}) where {BBS <: BBSearch} = Base.SizeUnknown()
 
 function iterate(search::BBSearch{DATA},
                  wt::BBTree=BBTree(root_element(search))) where {DATA}
-    isempty(wt.working_leafs) && return nothing
+
+    isempty(wt.working_leaves) && return nothing
 
     id = get_leaf_id!(search, wt)
-    X = wt[id]
+    X = wt.leaves[id]
     action, newdata = process(search, data(X))
     if action == :store
-        wt[id] = BBLeaf(newdata, parent_id(X), :final)
+        wt.leaves[id] = BBLeaf(newdata, X.parent, :final)
     elseif action == :bisect
-        parent = wt[id]
         child1, child2 = bisect(search, newdata)
         leaf1 = BBLeaf(child1, id, :working)
         leaf2 = BBLeaf(child2, id, :working)
         id1 = insert_leaf!(search, wt, leaf1)
         id2 = insert_leaf!(search, wt, leaf2)
-        wt[id] = BBNode(parent, id1, id2)
+        wt.nodes[id] = BBNode(X, id1, id2)
+        delete!(wt.leaves, id)
     elseif action == :discard
         discard_leaf!(wt, id)
     else
-        error("Branch and bound: process function of the search object return unknown action: $action, element $X is ignored. Valid actions are :store, :bisect and :discard.")
+        error("Branch and bound: process function of the search object return unknown action: $action for element $X. Valid actions are :store, :bisect and :discard.")
     end
     return wt, wt
 end