--- a/src/contractors.jl
+++ b/src/contractors.jl
@@ -1,3 +1,4 @@
+export Contractor
 export Bisection, Newton, Krawczyk
 
 """