--- a/src/roots.jl
+++ b/src/roots.jl
@@ -8,21 +8,21 @@ export start, next, done, copy, step!, eltype, iteratorsize
 diam(x::Root) = diam(x.interval)
 
 """
-    SearchStrategy(CONTAINER, store!, retrieve!)
+    SearchStrategy(CON, store!, retrieve!)
 
 Type describing the chosen strategy determining the order in which the
-intervals are processed. Given a type `EL`, the `CONTAINER` type must allow to
-create an empty container for it with the syntax `CONTAINER{EL}().
+intervals are processed. Given a type `EL`, the `CON` type must allow to
+create an empty container for it with the syntax `CON{EL}().
 The function `store!(container, element)` must add `element` to `container` and
 `retrieve!(container)` must return the next element to be processed and delete
 it from `set`.
 """
-struct SearchStrategy{CONTAINER}
+struct SearchStrategy{CON}
     store!::Function
     retrieve!::Function
 end
 
-SearchStrategy(CONTAINER::Type, store!, retrieve!) = SearchStrategy{CONTAINER}(store!, retrieve!)
+SearchStrategy(CON::Type, store!::Function, retrieve!::Function) = SearchStrategy{CON}(store!, retrieve!)
 
 SearchStrategy() = SearchStrategy(Vector, push!, pop!)
 
@@ -34,10 +34,10 @@ Returns the `RootSearchState` at each iteration. Note: Each iteration mutates
 the `RootSearchState`. Use `copy(state::RootSearchState)` to create an
 independent instance if necessary.
 """
-struct RootSearch{R <: Union{Interval,IntervalBox}, C <: Contractor, S <: SearchStrategy, T <: Real}
+struct RootSearch{R <: Union{Interval,IntervalBox}, C <: Contractor, CON, T <: Real}
     region::R
     contractor::C
-    strategy::S
+    strategy::SearchStrategy{CON}
     tolerance::T
 end
 
@@ -45,22 +45,22 @@ function RootSearch(region::R, contractor::C, tol::T) where {R <: Union{Interval
     RootSearch(region, contractor, SearchStrategy(), tol)
 end
 
-eltype(::Type{RS}) where {R, C, T, CONTAINER, S <: SearchStrategy{CONTAINER}, RS <: RootSearch{R, C, S, T}} = RootSearchState{CONTAINER{R}, CONTAINER{Root{R}}}
+eltype(::Type{RS}) where {R, C, T, CON, RS <: RootSearch{R, C, CON, T}} = RootSearchState{CON{R}, CON{Root{R}}}
 iteratorsize(::Type{RS}) where {RS <: RootSearch} = Base.SizeUnknown()
 
 
 struct RootSearchState{V, VR}
-    working::V  # Should be a container of the form  CONT{T}
-    outputs::VR  # Should ba a container of root of the form CONT{Root{T}}
+    working::V  # Should be a container of the form  CON{T}
+    outputs::VR  # Should ba a container of root of the form CON{Root{T}}
 end
 
 function RootSearchState(rs::RootSearch)
     return RootSearchState(rs.region, rs.strategy)
 end
 
-function RootSearchState(region::R, strat::S) where {R <: Union{Interval,IntervalBox}, CONTAINER, S <: SearchStrategy{CONTAINER}}
-    working = CONTAINER{R}()
-    outputs = CONTAINER{Root{R}}()
+function RootSearchState(region::R, strat::SearchStrategy{CON}) where {R <: Union{Interval,IntervalBox}, CON}
+    working = CON{R}()
+    outputs = CON{Root{R}}()
     strat.store!(working, region)
     return RootSearchState(working, outputs)
 end
@@ -79,7 +79,7 @@ copy(state::RootSearchState) =
     RootSearchState(deepcopy(state.working), deepcopy(state.outputs))
 
 
-function start(iter::RootSearch{R, C, S, T}) where {R, C, T, CONTAINER, S <: SearchStrategy{CONTAINER}}
+function start(iter::RootSearch{R, C, CON, T}) where {R, C, CON, T}
     state = RootSearchState(iter)
     return state
 end