… .travis.yml
… appveyor.yml
… examples/root_search_iterator.jl
… examples/roots.jl
… src/rootsearch_iterator.jl
… test/branch_and_bound.jl
… test/roots.jl
… test/runtests.jl
… test/search_interface.jl