I finally found the courage to update to 0.7/1.0 and to implement the idea proposed by @dpsanders  in #72 to separate the logic of the search from the rest of the package. Therefore this PR superseed #72 (I do not update it though, as the scope is quite different).

Changes in this PR:
- Introduction of a generic customizable branch and bound search interface.
- Introduction of a tree type as the fundamental data structure for the search. Some advantage come from that as for example pretty printing of the tree. The `roots` function still return a simple list of intervals.
- Separation of the branch and bound logic in the `rootsearch_iterator.jl` file, so the transition to the code in another package can be done easily.
- Rework of the interface to avoid storing functions in an object (and the kind of akwardly high number of parameters that resulted).
- Use of `Root` object during the search rather than interval, which is forced by the genericity of the branch and bound algorithm (status of roots does not exactly match with the status of the elements in a search).
- Fix of #29 along the way.