@dpsanders 2018-09-29
This looks great, thanks a lot!

I am a bit worried about the performance implications of using a tree like that. Have you done any benchmarks? 

In general, I think we should try to leave as general as possible the data structures used, so you could specify that you wanted to store the nodes in a tree, or a vector as before, etc. How possible would that be?