@Kolaru 2018-09-29
So I quickly added some benchmark using the Smiley and Chun examples introduced by @gwater and the `PkgBenchmark.jl` package ([this PR](https://github.com/JuliaCI/PkgBenchmark.jl/pull/68) is necessary to run its basics tasks though). Here are the results:

Master:

| ID                                                              | time            | GC time   | memory          | allocations |
|-----------------------------------------------------------------|----------------:|----------:|----------------:|------------:|
| `["Smiley", "Krawczyk", "Smiley and Chun (2001), Example 2.2"]` |  12.457 ms (5%) |           |   8.41 MiB (1%) |      250556 |
| `["Smiley", "Krawczyk", "Smiley and Chun (2001), Example 5.2"]` | 215.253 ms (5%) | 24.078 ms | 115.33 MiB (1%) |     3433780 |
| `["Smiley", "Krawczyk", "Smiley and Chun (2001), Example 5.4"]` |   9.902 ms (5%) |           |   7.51 MiB (1%) |      219406 |
| `["Smiley", "Newton", "Smiley and Chun (2001), Example 2.2"]`   |  12.983 ms (5%) |           |   8.76 MiB (1%) |      259155 |
| `["Smiley", "Newton", "Smiley and Chun (2001), Example 5.2"]`   | 255.685 ms (5%) | 28.522 ms | 136.83 MiB (1%) |     4044788 |
| `["Smiley", "Newton", "Smiley and Chun (2001), Example 5.4"]`   |  12.438 ms (5%) |           |   9.32 MiB (1%) |      271491 |

This PR:

| ID                                                              | time            | GC time   | memory          | allocations |
|-----------------------------------------------------------------|----------------:|----------:|----------------:|------------:|
| `["Smiley", "Krawczyk", "Smiley and Chun (2001), Example 2.2"]` |  12.540 ms (5%) |           |   8.46 MiB (1%) |      252695 |
| `["Smiley", "Krawczyk", "Smiley and Chun (2001), Example 5.2"]` | 226.340 ms (5%) | 27.452 ms | 116.05 MiB (1%) |     3450366 |
| `["Smiley", "Krawczyk", "Smiley and Chun (2001), Example 5.4"]` |   9.968 ms (5%) |           |   7.52 MiB (1%) |      220525 |
| `["Smiley", "Newton", "Smiley and Chun (2001), Example 2.2"]`   |  13.686 ms (5%) |           |   8.81 MiB (1%) |      261324 |
| `["Smiley", "Newton", "Smiley and Chun (2001), Example 5.2"]`   | 285.322 ms (5%) | 37.126 ms | 137.49 MiB (1%) |     4064908 |
| `["Smiley", "Newton", "Smiley and Chun (2001), Example 5.4"]`   |  12.503 ms (5%) |           |   9.35 MiB (1%) |      272998 |

It looks like this PR implies about 5% to 10% penalty (`PkgBenchmark` comparison is not yet functional for Julia 0.7/1.0, so it's the best estimate I can easily do for now). It may be possible to remove this penalty completely, since the tree structure closely correspond to a branch and bound algorithm.

On the other hand I was expecting way more allocations with this PR. Good surprise I guess.

As for allowing to store the data in arbitrary structure, while designin this PR I considerd that the ability to easily convert the result toward any data structure would be sufficient. It is also the simplest solution from the developper side.

However if the current penalty we pay to keep maximal information about the run of the algorithm is too high (seems reasonnable to me, but I essentially use the package to solve rather simple equations so I may not be representative), it is possible to make it more generic (may become tricky thought).