@Kolaru 2018-10-02
I updated the example in the `examples` dir and put it in a new file (`root_search_iterator.jl`). It is abundantly commented, so I hope everything is clear.

By default, the intermediate results are not stored in the tree to spare memory. However, one of the example show how to do this by changing the `process` function.

Also the search object already store all final parameters (final in the sense that the contractor is stored, but not the input parameter such as the original function or if it uses automatic differentiation for example).